@@ -1,41 +1,44 @@
 {
-  "name": "my-vue-quasar-app",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "run-p type-check \"build-only {@}\" --",
-    "preview": "vite preview",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --build",
-    "lint": "eslint . --fix",
-    "format": "prettier --write src/"
-  },
-  "dependencies": {
-    "@quasar/extras": "^1.16.14",
-    "pinia": "^2.2.6",
-    "quasar": "^2.17.4",
-    "swiper": "^11.1.15",
-    "vue": "^3.5.13",
-    "vue-router": "^4.4.5"
-  },
-  "devDependencies": {
-    "@quasar/vite-plugin": "^1.8.1",
-    "@tsconfig/node22": "^22.0.0",
-    "@types/node": "^22.9.3",
-    "@vitejs/plugin-vue": "^5.2.1",
-    "@vue/eslint-config-prettier": "^10.1.0",
-    "@vue/eslint-config-typescript": "^14.1.3",
-    "@vue/tsconfig": "^0.7.0",
-    "eslint": "^9.14.0",
-    "eslint-plugin-vue": "^9.30.0",
-    "npm-run-all2": "^7.0.1",
-    "prettier": "^3.3.3",
-    "sass": "^1.81.0",
-    "typescript": "~5.6.3",
-    "vite": "^6.0.1",
-    "vite-plugin-vue-devtools": "^7.6.5",
-    "vue-tsc": "^2.1.10"
-  }
+    "name": "my-vue-quasar-app",
+    "version": "0.0.0",
+    "private": true,
+    "type": "module",
+    "scripts": {
+        "dev": "vite",
+        "build": "run-p type-check \"build-only {@}\" --",
+        "preview": "vite preview",
+        "build-only": "vite build",
+        "type-check": "vue-tsc --build",
+        "lint": "eslint . --fix",
+        "format": "prettier --write src/"
+    },
+    "dependencies": {
+        "@quasar/extras": "^1.16.14",
+        "autoprefixer": "^10.4.20",
+        "pinia": "^2.2.6",
+        "postcss": "^8.4.49",
+        "quasar": "^2.17.4",
+        "swiper": "^11.1.15",
+        "vue": "^3.5.13",
+        "vue-router": "^4.4.5"
+    },
+    "devDependencies": {
+        "@quasar/vite-plugin": "^1.8.1",
+        "@tsconfig/node22": "^22.0.0",
+        "@types/node": "^22.9.3",
+        "@vitejs/plugin-vue": "^5.2.1",
+        "@vue/eslint-config-prettier": "^10.1.0",
+        "@vue/eslint-config-typescript": "^14.1.3",
+        "@vue/tsconfig": "^0.7.0",
+        "eslint": "^9.16.0",
+        "eslint-plugin-vue": "^9.32.0",
+        "npm-run-all2": "^7.0.1",
+        "prettier": "^3.3.3",
+        "sass": "^1.81.0",
+        "tailwindcss": "^3.4.15",
+        "typescript": "~5.6.3",
+        "vite": "^5.4.11",
+        "vite-plugin-vue-devtools": "^7.6.5",
+        "vue-tsc": "^2.1.10"
+    }
 }
