@@ -1,8 +1,8 @@
 declare module '@vue/runtime-core' {
-  export interface GlobalComponents {
-    RouterLink: typeof import('vue-router')['RouterLink']
-    RouterView: typeof import('vue-router')['RouterView']
-  }
+    export interface GlobalComponents {
+        RouterLink: (typeof import('vue-router'))['RouterLink']
+        RouterView: (typeof import('vue-router'))['RouterView']
+    }
 }
 
-export {} 
+export {}
