@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import './style/main.css'
 import { createPinia } from 'pinia'
+import './style/tailwind.css'
 
 import App from './App.vue'
 import router from './router'
@@ -13,7 +14,8 @@ const app = createApp(App)
 
 setupQuasar(app)
 
-app.use(createPinia())
+const pinia = createPinia()
+app.use(pinia)
 app.use(router)
 
 app.mount('#app')
