@@ -11,19 +11,19 @@ import langZhTw from 'quasar/lang/zh-TW'
 MaterialIconSet.table.arrowUp = 'sort'
 
 export const setupQuasar = (app: ReturnType<typeof createApp>) => {
-  app.use(Quasar, {
-    plugins: { Dialog, Notify, Loading, LoadingBar },
-    iconSet: MaterialIconSet,
-    config: {
-      brand: {
-        primary: 'var(--q-primary)',
-        secondary: 'var(--q-secondary)',
-        midd: 'var(--q-midd)',
-      },
-      loadingBar: {
-        color: 'primary',
-      },
-    },
-    lang: langZhTw,
-  })
+    app.use(Quasar, {
+        plugins: { Dialog, Notify, Loading, LoadingBar },
+        iconSet: MaterialIconSet,
+        config: {
+            brand: {
+                primary: 'var(--q-primary)', // 使用 CSS 變量
+                secondary: 'var(--q-secondary)',
+                midd: 'var(--q-midd)',
+            },
+            loadingBar: {
+                color: 'primary',
+            },
+        },
+        lang: langZhTw,
+    })
 }
