@@ -1,8 +1,8 @@
 export const isAuthenticated = (): boolean => {
-  return !!localStorage.getItem('token')
+    return !!localStorage.getItem('token')
 }
 
 export const logout = (): void => {
-  localStorage.removeItem('token')
-  window.location.href = '/login'
-} 
+    localStorage.removeItem('token')
+    window.location.href = '/login'
+}
