@@ -1,8 +1,13 @@
 <template>
-  <div class="about">
-    <h1 class="text-3xl font-bold">Hello world About!</h1>
-    <q-btn class="py-10" color="secondary" label="按鈕測試"></q-btn>
-  </div>
+    <div class="about">
+        <h1 class="text-3xl font-bold">Hello world About!</h1>
+        <q-btn class="py-10" color="secondary" label="按鈕測試"></q-btn>
+    </div>
 </template>
 <script lang="ts" setup></script>
-<style lang="scss"></style>
+<style lang="scss">
+.about {
+    width: 100%;
+    height: 100%;
+}
+</style>
