@@ -1,8 +1,4 @@
 <template>
-  <div class="about">
-    <h1 class="text-3xl font-bold">Hello world Home!</h1>
-    <q-btn class="py-10" color="secondary" label="按鈕測試"></q-btn>
-  </div>
+    <div class="bg-red-500 text-white p-4">如果你看到紅色背景，這說明 Tailwind CSS 已生效。</div>
 </template>
 <script lang="ts" setup></script>
-<style lang="scss"></style>
